hotel/src/requirements.txt
bcrypt==3.2.0
bottle==0.12.19
certifi==2021.5.30
cffi==1.14.4
charset-normalizer==2.0.4
cryptography==3.3.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==3.2
paramiko==2.7.2
phonenumbers==8.12.12
pycparser==2.20
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.26.0
rethinkdb==2.4.8
six==1.15.0
urllib3==1.26.6
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.2.0
tinydb
requests
